게임 화면 진행 상황 시간 정답확인과정 8.9(화)
게임다시만들어야함

+++ app/src/main/java/test/testapp/SpriteAnimation.java
@@ -36,7 +36,7 @@ public class SpriteAnimation extends GraphicObject {
 
     @Override
     public void Draw(Canvas canvas) {
-        Rect dest = new Rect(m_x, m_y, m_x + mSpriteWidth, m_y + mSpriteHeight);
+        Rect dest = new Rect((int)m_x, (int)m_y, (int) m_x + mSpriteWidth, (int) m_y + mSpriteHeight);
         canvas.drawBitmap(m_bitmap, mSRectangle, dest, null);
     }
